Refactor RP.RTC.Delay_Until to a separate package.
The implementation of Delay_Until depends on the ZFP interrupt
handler to disable the RTC interrupt when the alarm is triggered.
This is not compatible with the tasking runtimes where interrupts
are handled separately and the user may choose to install their
own interrupt handler via a protected object.

Delay_Until is therefore refactored into a separate package that
is only available in ZFP configurations.