RP.DMA: fix IRQ1 handling
There is a reserved word between the INTS0 and INTE1 which is not
represented in the current DMA_IRQs array. As a result, all register
accesses for IRQ1 are incorrect.

Adding a field to the DMA_IRQ record, either at the beginning or the
end, would overlap one the registers surrounding the DMA_IRQs array
(INTR or TIMER). The solution used here is to split the array in two
separate fields.